chore: rename contract button component
